requirements.txt
billiard
celery
redis
psycopg2-binary
python-dotenv